rearrange dmatrix modules a little
